+++ src/main/java/gift/config/LoginUserArgumentResolver.java
@@ -1,7 +1,7 @@
 package gift.config;
 
 import gift.auth.jwt.JwtProvider;
-import gift.domain.user.dao.UserJpaRepository;
+import gift.domain.user.repository.UserJpaRepository;
 import gift.domain.user.entity.User;
 import gift.exception.InvalidAuthException;
 import io.jsonwebtoken.Claims;
