+++ src/main/java/gift/domain/wishlist/service/WishlistService.java
@@ -1,9 +1,9 @@
 package gift.domain.wishlist.service;
 
-import gift.domain.product.dao.ProductJpaRepository;
+import gift.domain.product.repository.ProductJpaRepository;
 import gift.domain.product.entity.Product;
 import gift.domain.user.entity.User;
-import gift.domain.wishlist.dao.WishlistJpaRepository;
+import gift.domain.wishlist.repository.WishlistJpaRepository;
 import gift.domain.wishlist.dto.WishItemDto;
 import gift.domain.wishlist.entity.WishItem;
 import gift.exception.InvalidProductInfoException;
