+++ src/test/java/gift/domain/user/service/UserServiceTest.java
@@ -8,7 +8,7 @@ import static org.mockito.Mockito.mock;
 
 import gift.auth.jwt.JwtProvider;
 import gift.auth.jwt.Token;
-import gift.domain.user.dao.UserJpaRepository;
+import gift.domain.user.repository.UserJpaRepository;
 import gift.domain.user.dto.UserDto;
 import gift.domain.user.dto.UserLoginDto;
 import gift.domain.user.entity.Role;
